feat: protezione rotte
