+++ app/account/page.tsx
@@ -1,5 +1,6 @@
 import { createClient } from "utils/supabase/server";
 import AccountForm from "./account-form";
+import { redirect } from "next/navigation";
 
 export default async function Account() {
   const supabase = createClient();
@@ -7,6 +8,10 @@ export default async function Account() {
   const {
     data: { user },
   } = await supabase.auth.getUser();
+
+  if (!user) {
+    return redirect("/login");
+  }
 
   return <AccountForm user={user} />;
 }
